[tests] separate test framework into separate module
